+++ R/utils.R
@@ -314,6 +314,19 @@ named <- function(x) {
   x
 }
 
+# Copied from pillar.
+#
+# Function for the thousand separator, returns "," unless it's used for the
+# decimal point, in which case returns "."
+big_mark <- function(x) {
+  # The thousand separator,
+  # "," unless it's used for the decimal point, in which case "."
+  mark <- if (identical(getOption("OutDec"), ",")) "." else ","
+  ret <- formatC(x, big.mark = mark, format = "d", preserve.width = "individual")
+  ret[is.na(x)] <- "??"
+  ret
+}
+
 browser <- function(...,
                     skipCalls = 0,
                     frame = parent.frame()) {